fix: widen test cfg_addr variables to uint64_t
  Use uint64_t for local cfg_addr variables in the affected DA and
  DPT tests.

  This avoids truncation of PCIe config-space addresses returned by
  val_pcie_get_bdf_config_addr() on platforms using high address
  ranges.

+++ test_pool/da/da_selective_ide_register_property.c
@@ -67,7 +67,7 @@ payload(void)
   uint32_t pgt_attr_el3;
   uint32_t sel_str_lock_bit;
   uint64_t va;
-  uint32_t cfg_addr;
+  uint64_t cfg_addr;
   uint32_t da_cap_base;
   uint32_t stream_id;
   uint32_t str_index;
